refactor: delete components and relative unit test
