+++ src/__tests__/components/EditorRightSideBar/ColorBoard.test.tsx
@@ -1,6 +1,6 @@
 import { act, screen } from "@testing-library/react";
 import user from "@testing-library/user-event";
-import ColorBoard from "@/components/TopBar/ColorBoard";
+import ColorBoard from "@/components/EditorRightSideBar/ColorBoard";
 import renderWithMockedProvider from "../../utils";
 import store from "@/store/index";
 import { initialState, PriceBar } from "@/store/reducer/priceBarSlice";
