+++ src/layouts/index.tsx
@@ -4,9 +4,6 @@ import { useRouter } from "next/router";
 import Header from "./Header";
 import NavigationBar from "../components/NavBar";
 import EditorSideBar from "../components/EditorSideBar";
-import EditorFooter from "@/components/EditorFooter";
-import PriceBar from "@/components/PriceBar";
-import TopBar from "@/components/TopBar";
 import ShoppingCart from "@/components/ShoppingCart";
 import MyTemplate from "@/components/MyTemplate";
 import { PAGE_NOT_FOUND, PAYMENT, MY_ORDER } from "../../src/constants";
@@ -63,12 +60,9 @@ const Layout: React.FC<{ children: ReactNode }> = ({ children }) => {
             <OrderGeneration />
             <EditorSideBar />
             {children}
-            <TopBar />
             <EditorRightSideBar />
           </Box>
           <TileBoard />
-          <PriceBar />
-          <EditorFooter />
         </>
       );
   }
